refactor: WebViewController 한 클래스로 통합
