--- src/main/java/cart/controller/web/CartController.java
@@ -1,23 +1,0 @@
-package cart.controller.web;
-
-import cart.service.CartService;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/cart")
-public class CartController {
-
-    private final CartService cartService;
-
-    public CartController(final CartService cartService) {
-        this.cartService = cartService;
-    }
-
-
-    @GetMapping
-    public String renderCartPage() {
-        return "cart";
-    }
-}
